refactor(pw-strength): Remove designF's "experiment" status.
DesignF is no longer considered experimental, it's the default.

Because experiments are not enabled by default in functional
tests, the functional tests aren't testing reality. The
pw-strength UI is not displayed in the functional tests unless
the magic query params `&forceExperiment=passwordStrength&passwordExperimentGroup=designF`
are added to the URL.

Note, password strength experiment metrics will no longer be
sent to Amplitude.

blocks #6572
blocks #6573
fixes #6564
fixes #6400
