+++ app/scripts/templates/sign_up_password.mustache
@@ -21,21 +21,11 @@
 
       <div class="input-row password-row">
         <input id="password" type="password" class="password check-password tooltip-below" placeholder="{{#t}}Password{{/t}}" value="{{ password }}" pattern=".{8,}" required autofocus data-form-prefill="password" data-synchronize-show="true" />
-        {{#showCustomHelperBalloon}}
-          <div class="helper-balloon"></div>
-        {{/showCustomHelperBalloon}}
+        <div class="helper-balloon"></div>
       </div>
 
       <div class="input-row password-row">
         <input id="vpassword" type="password" class="password check-password tooltip-below" placeholder="{{#t}}Repeat password{{/t}}" pattern=".{8,}" required data-synchronize-show="true" />
-        {{^showCustomHelperBalloon}}
-        <div class="input-help input-help-focused input-help-signup input-help-small-only">{{#t}}Don’t lose this password. If you need to reset it, you may lose Sync data.{{/t}}</div>
-        <div class="input-help input-help-focused input-help-signup input-help-big-only input-help-balloon follow-focus">
-          {{#unsafeTranslate}}<span>Don’t reuse your email password</span>If someone knows your email password, they can approve your Firefox sign-in on a new device and access your data.{{/unsafeTranslate}}
-          <br/><br/>
-          {{#unsafeTranslate}}<span>Remember your password</span>If you need to reset your password, you will lose any Sync data that is not on one of your devices.{{/unsafeTranslate}}
-        </div>
-        {{/showCustomHelperBalloon}}
       </div>
 
       {{{ coppaHTML }}}
