+++ app/tests/spec/lib/experiments/grouping-rules/index.js
@@ -12,7 +12,7 @@ define(function (require, exports, module) {
 
   describe('lib/experiments/grouping-rules/index', () => {
     it('EXPERIMENT_NAMES is exported', () => {
-      assert.lengthOf(ExperimentGroupingRules.EXPERIMENT_NAMES, 9);
+      assert.lengthOf(ExperimentGroupingRules.EXPERIMENT_NAMES, 8);
     });
 
     describe('choose', () => {
