+++ app/tests/test_start.js
@@ -42,13 +42,11 @@ require('./spec/lib/environment');
 require('./spec/lib/error-utils');
 require('./spec/lib/experiment');
 require('./spec/lib/experiments/base');
-require('./spec/lib/experiments/password-strength');
 require('./spec/lib/experiments/grouping-rules/base');
 require('./spec/lib/experiments/grouping-rules/communication-prefs');
 require('./spec/lib/experiments/grouping-rules/email-first');
 require('./spec/lib/experiments/grouping-rules/index');
 require('./spec/lib/experiments/grouping-rules/is-sampled-user');
-require('./spec/lib/experiments/grouping-rules/password-strength');
 require('./spec/lib/experiments/grouping-rules/q3-form-changes');
 require('./spec/lib/experiments/grouping-rules/recovery-key');
 require('./spec/lib/experiments/grouping-rules/send-sms-install-link');
@@ -185,7 +183,7 @@ require('./spec/views/mixins/one-visible-of-type-mixin');
 require('./spec/views/mixins/open-webmail-mixin');
 require('./spec/views/mixins/password-mixin');
 require('./spec/views/mixins/password-reset-mixin');
-require('./spec/views/mixins/password-strength-experiment-mixin');
+require('./spec/views/mixins/password-strength-mixin');
 require('./spec/views/mixins/pulse-graphic-mixin');
 require('./spec/views/mixins/resend-mixin');
 require('./spec/views/mixins/resume-token-mixin');
