+++ tests/functional.js
@@ -7,7 +7,6 @@ module.exports = [
   'tests/functional/mailcheck.js',
   'tests/functional/sync_v3_email_first.js',
   'tests/functional/fx_firstrun_v2_email_first.js',
-  'tests/functional/password_strength_experiment.js',
   // new and flaky tests above here',
   'tests/functional/sign_in.js',
   'tests/functional/sign_in_cached.js',
@@ -70,6 +69,7 @@ module.exports = [
   'tests/functional/pages.js',
   'tests/functional/back_button_after_start.js',
   'tests/functional/cookies_disabled.js',
+  'tests/functional/password_strength.js',
   'tests/functional/password_visibility.js',
   'tests/functional/avatar.js',
   'tests/functional/alternative_styles.js',
