+++ tests/functional/send_sms.js
@@ -35,6 +35,7 @@ let formattedPhoneNumber;
 
 const {
   click,
+  clearBrowserState,
   closeCurrentWindow,
   deleteAllSms,
   disableInProd,
@@ -73,6 +74,7 @@ const suite = {
     // User needs a sessionToken to be able to send an SMS. Sign up,
     // no need to verify.
     return this.remote
+      .then(clearBrowserState({ force: true }))
       .then(fillOutSignUp(email, PASSWORD))
       .then(testElementExists(selectors.CONFIRM_SIGNUP.HEADER))
       // The phoneNumber can be reused by different tests, delete all
